+++ internal/core/domain/leasing_history.go
@@ -13,12 +13,11 @@ type LeasingHistory struct {
 	Dorm     Dorm      `gorm:"foreignKey:DormID;references:ID"`
 	LesseeID uuid.UUID `gorm:"type:uuid;not null"`
 	Lessee   User      `gorm:"foreignKey:LesseeID;references:ID"`
-	ReviewID uuid.UUID `gorm:"type:uuid"`
-	Review   Review    `gorm:"foreignKey:ReviewID;references:ID"`
 	Orders   []Order   `gorm:"foreignKey:LeasingHistoryID"`
 	Start    time.Time
 	End      time.Time `gorm:"default:null"`
 	Price    float64
+	Review   Review `gorm:"embedded"`
 }
 
 func (l *LeasingHistory) ToDTO() dto.LeasingHistory {
@@ -35,5 +34,19 @@ func (l *LeasingHistory) ToDTO() dto.LeasingHistory {
 		Start:  l.Start,
 		End:    l.End,
 		Price:  l.Price,
+	}
+}
+
+type Review struct {
+	Message  string
+	Rate     int
+	CreateAt *time.Time `gorm:"autoCreateTime"`
+}
+
+func (r *Review) ToDTO() dto.Review {
+	return dto.Review{
+		Message:  r.Message,
+		Rate:     r.Rate,
+		CreateAt: *r.CreateAt,
 	}
 }
